+++ framework/changelog.md
@@ -1,6 +1,5 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- Upgrade dependency: core to 1.2.8
-- Chore: Moves all context key types to schemas.BifrostContextKey
-- Chore: Adds new logs table migration to avoid missing any required columns in the DB
+- Upgrade dependency: core to 1.2.9
+- Fix: JSON serialization for error objects and tool function parameters