+++ plugins/jsonparser/changelog.md
@@ -1,5 +1,4 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- Upgrade dependency: core to 1.2.8
-- Chore: Moves all context key types to schemas.BifrostContextKey
+- chore: version update core to 1.2.9